🔄 Refactor catdepth_new import: move to S_Category directory and update references in page modules
