+++ ncc_page.py
@@ -34,7 +34,7 @@ import configparser
 from pathlib import Path
 from newapi.super import bot_api
 from newapi.super import super_page
-from newapi.super import catdepth_new
+from newapi.super.S_Category import catdepth_new
 
 print_test = {1: True if "test" in sys.argv else False}
