+++ page.py
@@ -28,7 +28,7 @@ import os
 import sys
 from newapi.super import bot_api
 from newapi.super import super_page
-from newapi.super import catdepth_new
+from newapi.super.S_Category import catdepth_new
 
 from newapi import useraccount
