+++ super/S_Category/catdepth_new.py
