+++ wiki_page.py
@@ -29,7 +29,7 @@ import os
 # import sys
 from newapi.super import bot_api as bot_api1
 from newapi.super import super_page as super_page1
-from newapi.super import catdepth_new as catdepth_new1
+from newapi.super.S_Category import catdepth_new as catdepth_new1
 
 from newapi import user_account_new
